+++ NEWS.md
@@ -4,6 +4,9 @@
 
 * `format_bf()` gains a digits argument.
 
+* More informative error message when the response variable in `get_response()`
+  is not present in the data.
+
 # insight 1.5.0
 
 ## Changes